move Wumpex.Gateway.Worker to Wumpex.Gateway
Refactor logic of the worker into the Gateway module
Updated Shard to use the new Gateway instead of Worker
