+++ lib/wumpex/shard.ex
@@ -78,7 +78,7 @@ defmodule Wumpex.Shard do
     {:ok, _websocket} =
       DynamicSupervisor.start_child(
         supervisor,
-        {Wumpex.Gateway.Worker, [
+        {Wumpex.Gateway, [
           host: gateway, port: 443, path: "/?v=8&encoding=etf", timeout: 5_000,
           shard: shard, gateway: gateway, token: Wumpex.token(), guild_sup: guild_sup
         ]}
